refactor: restructure
