+++ .vscode/launch.json
@@ -10,7 +10,7 @@
             "request": "launch",
             "module": "flask",
             "env": {
-                "FLASK_APP": "apps/app.py",
+                "FLASK_APP": "src/diet/app.py",
                 "FLASK_DEBUG": "1"
             },
             "args": [
