+++ src/diet/auth/repository.py
@@ -1,8 +1,8 @@
 from sqlalchemy.exc import SQLAlchemyError
 
-from apps.auth.models import User
-from apps.extensions import sql_alchemy
-from apps.utils.custom_logger import CustomLogger
+from diet.auth.models import User
+from diet.extensions import sql_alchemy
+from diet.utils.custom_logger import CustomLogger
 
 _logger = CustomLogger.get_logger()
