+++ src/diet/body_composition/templates/body_composition/record_body_composition.html
