+++ src/diet/static/color-modes.js
