+++ src/diet/templates/404.html
