+++ src/diet/templates/500.html
