+++ src/diet/utils/directory_util.py
