+++ tests/diet/body_composition/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the body_composition package."""
