+++ tests/diet/body_composition/test_views.py
@@ -1,0 +1,15 @@
+class TestLogBodyComposition:
+    # 未ログインの場合は、ログインページに遷移すること。
+    # ログイン済みの場合は体重記録のページに遷移すること。
+    # ログイン済みの場合は入力欄の初期値が最新の情報になっていること。
+    # （日付は今日になっていればよかったはず。）
+    # 正しい値を入力した場合、登録されてホームページに遷移すること。
+    # 不正な値を入力した場合、エラーが表示されること。
+    # 登録済みの値を更新できること。
+    pass
+
+
+class TestFetchBodyCompositionData:
+    # 未ログインの場合はサンプルデータが返ってきていること。
+    # ログインしている場合は、そのユーザーの情報が返ってきていること。
+    pass
